#43214: refactor

+++ src/main/java/com/topface/topface/ui/adapters/BindingAdapterMethods.java
@@ -1,8 +1,11 @@
 package com.topface.topface.ui.adapters;
 
 import android.databinding.BindingAdapter;
+import android.databinding.ObservableFloat;
 import android.view.View;
 import android.widget.LinearLayout;
+
+import com.topface.framework.utils.Debug;
 
 /**
  * Created by ppetr on 14.01.16.
@@ -10,8 +13,8 @@ import android.widget.LinearLayout;
  */
 public class BindingAdapterMethods {
     @BindingAdapter("android:layout_marginTop")
-    public static void setMarginTop(View view, int padding) {
+    public static void setMarginTop(View view, float padding) {
         LinearLayout.LayoutParams lp = (LinearLayout.LayoutParams) view.getLayoutParams();
-        lp.setMargins(lp.leftMargin, padding, lp.rightMargin, lp.bottomMargin);
+        lp.setMargins(lp.leftMargin, (int) padding, lp.rightMargin, lp.bottomMargin);
     }
 }
